+++ src/text/tesseract.rs
@@ -1,19 +1,10 @@
 //! Tesseract
+use super::*;
+use super::private::*;
 use ::*;
-use core::*;
 use std::ffi::*;
 use std::os::raw::c_char;
 use std::path::Path;
-
-mod private {
-    #[allow(missing_copy_implementations, missing_debug_implementations)]
-    pub enum COCR {}
-
-    pub trait OcrImpl {
-        fn get_value(&self) -> *mut COCR;
-    }
-}
-
 
 extern "C" {
     fn cv_tesseract_new(
@@ -22,17 +13,8 @@ extern "C" {
         char_whitelist: *const c_char,
         oem: EngineMode,
         psmode: PageSegmentationMode,
-    ) -> *mut private::COCR;
-    fn cv_tesseract_drop(ocr: *mut private::COCR);
-    fn cv_ocr_run(
-        ocr: *const private::COCR,
-        image: *const CMat,
-        output_text: *mut CDisposableString,
-        component_rects: *mut CVec<Rect>,
-        component_texts: *mut CVec<CDisposableString>,
-        component_confidences: *mut CVec<f32>,
-        component_level: ComponentLevel,
-    );
+    ) -> *mut COCR;
+    fn cv_tesseract_drop(ocr: *mut COCR);
 }
 
 #[repr(C)]
@@ -70,19 +52,10 @@ pub enum PageSegmentationMode {
     SingleChar,
 }
 
-#[allow(missing_docs)]
-pub trait OcrImplInterface: private::OcrImpl {}
-
-/// Basic trait for all OCR types
-pub trait Ocr {
-    /// Recognize text
-    fn run(&self, image: &Mat, component_level: ComponentLevel) -> (String, Vec<Rect>, Vec<String>, Vec<f32>);
-}
-
 /// `OcrTesseract` class provides an interface with the tesseract-ocr API
 #[derive(Debug)]
 pub struct OcrTesseract {
-    value: *mut private::COCR,
+    value: *mut COCR,
 }
 
 impl OcrTesseract {
@@ -117,40 +90,13 @@ impl Drop for OcrTesseract {
     }
 }
 
-impl private::OcrImpl for OcrTesseract {
-    fn get_value(&self) -> *mut private::COCR {
+impl OcrImpl for OcrTesseract {
+    fn get_value(&self) -> *mut COCR {
         self.value
     }
 }
 
 impl OcrImplInterface for OcrTesseract {}
-
-impl<T: OcrImplInterface> Ocr for T {
-    fn run(&self, image: &Mat, component_level: ComponentLevel) -> (String, Vec<Rect>, Vec<String>, Vec<f32>) {
-        let value = self.get_value();
-        let mut output_text = CDisposableString::default();
-        let mut component_rects = CVec::<Rect>::default();
-        let mut component_texts = CVec::<CDisposableString>::default();
-        let mut component_confidences = CVec::<f32>::default();
-        unsafe {
-            cv_ocr_run(
-                value,
-                image.inner,
-                &mut output_text,
-                &mut component_rects,
-                &mut component_texts,
-                &mut component_confidences,
-                component_level,
-            );
-        }
-        (
-            output_text.unpack(),
-            component_rects.unpack(),
-            component_texts.unpack(),
-            component_confidences.unpack(),
-        )
-    }
-}
 
 fn to_nullable_string(value: &Option<CString>) -> *const c_char {
     unwrap_or_null(&value.as_ref().map(|x| x.as_ptr()))
